* Updated Cover image on First Section. * Added header Component. * Rename typhography for Heading (More semantically correct). * Added text component [Eyebrow, Body, Body Large].
