+++ src/components/common/header/header.scss
@@ -1,13 +1,18 @@
 @import '../../../scss/constant/header';
+@import '../../../scss/mixins/media-queries';
 
 .cyberpunk-header {
   align-items: center;
   display: flex;
   justify-content: space-between;
   left: 0;
-  padding: 40px 60px;
+  padding: 10px;
   position: fixed;
   top: 0;
   width: 100%;
   z-index: $cyberpunk-header-z-index;
+
+  @include bp(lg) {
+    padding: 40px 60px;
+  }
 }
