+++ src/components/common/logo/logo.scss
@@ -1,3 +1,9 @@
+@import '../../../scss/mixins/media-queries';
+
 .cyberpunk-logo {
-  max-height: 70px;
+  max-height: 40px;
+
+  @include bp(lg) {
+    max-height: 70px;
+  }
 }
