+++ src/components/ui-system/atoms/typography/heading.scss
@@ -1,17 +1,5 @@
 @import '../../../../scss/constant/colors';
 
-@font-face {
-  font-family: 'Blender Pro';
-  font-weight: 700;
-  src: url('/fonts/BlenderPro-Bold.ttf');
-}
-
-@font-face {
-  font-family: 'Blender Pro Book';
-  font-weight: 700;
-  src: url('/fonts/BlenderPro-Book.ttf');
-}
-
 $headline-weight: 700;
 
 .cyberpunk-headline {
