+++ src/components/ui-system/atoms/typography/text.scss
@@ -1,0 +1,7 @@
+@import '../../../../scss/constant/colors';
+
+$headline-weight: 700;
+
+.cyberpunk-eyebrow {
+  font-size: 16px;
+}
