+++ src/content/hero/hero.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Img from 'gatsby-image';
 import { useStaticQuery, graphql } from 'gatsby';
-import Heading from '../../components/ui-system/atoms/typography/typhography';
+import Heading from '../../components/ui-system/atoms/typography/heading';
 import content from '../../data/content/hero.json';
 import './hero.scss';
