+++ src/scss/constant/_breakpoints.scss
@@ -1,0 +1,4 @@
+$sm-device: 576px;
+$md-device: 768px;
+$lg-device: 992px;
+$xl-device: 1200px;
